Refine set syntax and more
The main changes are:

Removed the overloading of the equality operator to do set membership tests
and updated set generation syntax to align with the new membership test
syntax.

Also, got rid of the "[]" syntax in references and replaced with special "_"
character that is treated as a special iterator. Documented this behaviour in
the references section.

Also, fixed semantics so that document content is defined in terms of the rule
head which is in turn defined by instances of the body that evaluate to true.
E.g., instead of "p :- 7", we now write "p = 7", which is short for "p = 7 :-
true".