Add module and package support to REPL
The REPL now adds a module to the policy store when it initializes. This will
allow the REPL to represent configuration as policy later. In addition, the
REPL keeps track of the currently active module. The modules are identified by
the package name.

The package directive causes the REPL to change the currently active module.
This allows the user to effectively 'cd' into packages and reference docs in
that package without a prefix.

The import directive causes the REPL to add the imported package to the
currently active module.

Also, rename REPL state variables to make them private.