Fix bug in complete ref/global handling
Errors were being silently dropped when the result of reference evaluation was
an array or object AST type and a later expression encountered an error. The
Query function on AST types does not have to return an error if the input
reference is invalid for the specific AST value and because this makes error
handling more complicated, just remove it.