Error message improvements
These changes add some helpers to obtain contextual information for error
mesasages. Specifically, the Location values on all AST nodes can be used to
format error messages that include line, column, and filename.

Also, add compiler step to check that built-ins are provided with the correct
number of arguments. The built-in implementations in topdown assume this is
the case and will index into the term slice without checking.