Add REST API documentation
The REST API has been manually documented in
site/documentation/references/rest/index.md.

Each of the API endpoints is listed with:

- HTTP method and path
- Query parameters
- Response status codes
- Example request/response transactions

In the future, we could consider modelling the REST APIs using a framework
like Swagger, however, at this point in time, the existing Go-based solutions
either (1) require implementing the entire API inside of another framework
(e.g., go-restful) or (2) lack sufficient support for comment-based
annotations (e.g., go-swagger).

For now, given the relatively small size of the API, it should be OK to
manually update the docs as the API evolves.