Fix expression evaluation for refs to base docs
If the expression consists of a single ref to a base doc, the ref must be
looked up before the "truth check" to determine if evaluation continues.

In 34b1c86 we modified the evaluation behaviour to continue as long as
the term is not Boolean(false). In doing so, we forgot to handle the case
where a reference refers to a "false" value in storage.

This is confusing in cases where you have some reference to a flag on an
object, e.g., "..., data.servers[i].interfaces[j].enabled, ...".

This problem does not arise with references to virtual docs because the
reference would be bound in the context.

With this change, the REPL had to be updated, otherwise we would not display
"false" values when evaluating a non-ground reference. E.g., given data = {"x":
[false,true]}, data.x[i] should print 0/false, 1/true (not just 1/true).