Add basic transaction support in Storage
Transactions are added to the storage layer to facilitate consistent snapshots
over data provided by storage plugins. For now, the transactions are
serialized by the Storage object using a lock. If/when throughput becomes a
problem, this can be replaced with a more sophisticated locking scheme or
other concurrency control (hopefully) without requiring major changes to the
storage layer APIs.