Fix panic on unhashable key type
The cache that was added to the topdown implementation assumed rules
producing objects would always yield hashable/scalar keys. If the key was not
hashable, the cache lookup would cause a panic.

This change updates the topdown implementation to check the type of the object
key that was produced and return an intelligible error message if necessary.