Build optimizations for normal use and docker. Depends on #124
Since we have a static binary now, we don't have to build for different
OSs. Also, commit this binary to version control so that we can setup
automatic builds on hub.docker.com and don't have to have golang in the
docker image. Finally use alpine so that logging to stderr will work
(cause there is a FS in scratch), and to prepare for having this listen
over TLS. Image size grew 3.73MB which I think is reasonable for how
much it gives us over scratch.