Update AST to support comparisons on all types
This was necessitated by need to compare Sets, but it should be generally
useful to be able to apply sorting to AST elements. The sorting for objects
and sets as well as higher level constructs (e.g., expressions, rules, etc.)
are consistent but not defined. This follows the approach taken in Python.