Add new compiler harness for ad-hoc queries
This change adds a QueryCompiler interface and default implementation to the
ast package. QueryCompiler allows callers to compile ad-hoc queries in the
context of an existing compiler instance.

Previously, callers were forced to construct a fake module and rule to contain
the query for compilation. Then, the caller had to re-compile all modules.
This is unnecessarily expensive and also prevents callers from compiling the
obvious query "data" (because it would be recursive).

One downside of the new approach is that it requires a bit of extra
boilerplate to run the compilation on queries (in addition to modules). This
seems acceptable for now. As the core compiler logic can be factored into
separate places and shared.