Update topdown package with request changes
These changes replace the Globals map with a single Request value that
contains all of the supplied query arguments. This way it's still possible for
the caller to provide a single, non-object value as the request input.

Because request has the same semantics as data now, users will no longer
receive a special error if the request is missing a required key. This
behaviour was somewhat broken anyway as missing embedded keys would not
trigger the error. This seems better left to schema checking at the API layer.