Fix reference evaluation bug
References were not being evaluated correctly. The top-level evalRefRec
function was computing the ground prefix from the reference's binding
and then performing the rest of the evaluation with the original
reference and the prefix. Instead, the rest of the evaluation should
proceed with the plugged value and the prefix. This fix was not
compatible with how reference bindings were stored. Before, reference
bindings were keyed with the original (potentially non-ground)
reference. Now, they are keyed with the evaluated version of the
reference. This version is guaranteed to be ground. As a result, the
PlugValue handling of references must be updated to handle cases such
as:

orig ref:   p[x][y]
locals:     {x: 1, y: 2}
refs:       {p[1][2]: "foo"}

The solution is to recursively plug the ref. E.g., p[x][y] => p[1][y] =>
p[1][2] => "foo".

Fixes #238