Fix reordering for unsafe ref heads
Evaluation assumes that ref heads are safe when evaluating expressions.
In the future this might be relaxed, however for now, reordering must
not consider vars safe if they would unify with a ref that has an unsafe
head.

These changes also update the compiler to keep track of generated vars
in each module. This allows the compiler to suppress error messages for
unsafe generated vars--this is fine as an unsafe generated var message
will always be accompanied with another unsafe var message. It would be
confusing to users if unsafe var messages were reported for vars that
they had not specified.

Also, refactored safety check test case. It was becoming difficult to
track down errors because the test case exercised the entire module at
once. Now the individual test cases are isolated and it's easier to
identify errors.

Fixes #297