Improve file loader error handling
The file loader now handles empty modules properly (previously, it would
include an empty/nil module in the results which resulted in a panic
later on).

Also, the file loader now accumulates errors instead of bailing on the
first one. This makes it easier to find and fix errors when loading
files into OPA.