Add rule indexing to compiler
These changes add basic indexing support to the compiler. Rules can be
indexed based on their contents. In this case, the first indexing
strategy we support is based on expressions of the form: eq(term,
base_doc_ref) where the term doesn't require evaluation (except vars).
In these cases, we construct a trie that allows us to efficiently search
for rules that must be evaluated (rules that do not have to be evaluated
are excluded by the search.)

As part of these changes the virtual doc benchmark has been updated to
include rules that are hit and missed by the rule indexer.