Add io.jwt.decode builtin
It would be useful to support policies involving access tokens. To
facilitate that, io.jwt.decode takes a JSON Web Token in JWS Compact
Serialization representation as input and parses it into its
header, payload and signature.

This patch does its best to adhere to the following RFCs where
relevant:
- JWT: https://tools.ietf.org/html/rfc7519
- JWS: https://tools.ietf.org/html/rfc7515
- JWE: https://tools.ietf.org/html/rfc7516