Add else keyword support to rule indexing
These changes add support for the else keyword to the rule indexing. As
part of these changes, the index interface has been modified to
accomodate the chains of 'else' rules that may be attached to any
rule.

The trie traversal has been modified to record the priority of roots
that are hit in the lookup. This allow the index to return rules in a
deterministic order. This is required because the hits have to be
grouped by roots; if groups were stored in a map, iteration order would
become non-deterministic. This way we can construct the
IndexResult.Rules slice deterministically.