Update walker to call AST visit function on Term
Before, the walker would skip the Term and move straight to the Term
value. In some cases, it's useful to be able to process Terms
generically.

Note, this required removing the filename setter as some terms do not
have a location set. This is a separate issue that will be fixed in the
next commit.