Fix setting of term locations
There were a handful of places in the parser where the location was not
being set on terms. All parsed terms should have a location set (for
debug/error purposes.) These changes correct the missing cases in the
parser. Also, now that the visitor can process terms, we can add back
the filename setter on terms.