Implement basic notification service in the server
This patch adds server support for setting watches on base data
documents. When making a GET request, the user can specify that they
wish to watch the path or not with the `watch` query parameter. The
server will register their source address and send notifications to
it whenever the path is modified.