Improve type checking error messages
These changes improve the type checking error messages for undefined
refs. The improvements split the ref errors into different categories:

1) Unsupported ref errors - parent cannnot be dereferenced
2) Invalid ref errors - ref operand is invalid
3) Missing type info errors - ref head type information is missing

For unsupported ref errors, the type of the parent is reported and the
prefix is indicated with carets.

For invalid ref errors, the ref operand is reported and indicated with a
carrot.

Also, add coverage of check error details output.