Add ReadOne and WriteOne convenience functions to the storage layer
It's common for operations to the store to simply read or write a single
value. The ReadOne and WriteOne functions each create their own
transactions, perform the operation and then clean up after themselves.
This greatly simplifies user code that only performs singular
operations, as it no longer needs to check for errors at every step.