Allow partial docs to be defined without bodies in modules
Rego already supports syntax for declaring constants in modules,
parsing them as lone Bodies and then transforming them into rules
with a Body of `true`. This patch adds similar behavior for partially
defined object documents, allowing keys of the document to be defined
as module level constants.

As an example, instead of writing:

    package foo

    bar["x"] = 1 { true }
    bar["y"] = 2 { true }

One could omit the Bodies and just write:

    package foo

    bar["x"] = 1
    bar["y"] = 2

This is limited to a depth of one, that is, a constant like
`bar["x"]["y"] = 1` is not supported.