Remove panic workaround in parser
The new version of mna/pigeon fixes the location bug that prevented
accurate line/column numbers from being reported on unmatched input.

In the future we can investigate improvements to the error messages by
adding negative matches that return higher-level errors as long as the
location information is not impacted.

Fixes #214