Fix unify to return undos in all cases
The evalEqUnify implementation was not returning undo objects in cases
where it should: if unification failed after binding a var, the binding
was not undone and backtracking would proceed with an incorrect binding.
For this to manifest, a query would likely be iterating and performing
pattern matching on a composite value.

Fixes #436