Fix type inference bug
The type checker was starting from the head for the refs to base docs
and inferring type Any for the head var as well as any other vars found
in the ref. The checker should never have inferred type Any for the
head.

With this change in place, the server test of watch invalidation now
correctly reports a compile-time error.

Fixes #381