Fix file watch bug causing panic in server mode
Previously, the runtime file watcher would not include the policies in
storage when recompiling after a file watch update. As a result, if a
policy snuck into storage "somehow" then the server would panic when it
attempted to reload the policies.

A policy could sneak into storage in two cases...

1) File removed AFTER FS notification sent but BEFORE loader.All()
called by file watcher.

2) File watcher enabled and policies pushed into server via API.

In either case, the policies would need to conflict with each other such
that compilation would fail in the server (e.g., with a type error.)