Add basic out-of-band metrics collection
Collect Duration metrics form the http handlers. Instrument each handler
using the promhttp package. The metrics are partitioned on their status code,
the name of the handler serving the request and the HTTP method, as
"code", "handler" and "method" respectively.

To avoid unbounded cardinality an approximation of the request's url is
used as the "handler" dimension. The handlers roughly correspond to the
different api resource path & the api version.

Issue: #525

Signed-off-by: Ross Guarino <rssguar@gmail.com>