Fix safety check for nested function calls
Refactor how output vars are computed for call expressions. Previously
the number of rule args were not used to determine which args were
considered outputs. Instead, it was assumed the last arg in the call was
an output. This meant that if an arg in the output position was omitted,
an input arg would be incorrectly marked safe.

With these changes, the compiler looks up the number of args (arity) of
the rule when checking whether an arg is an output.