Improve partial eval of negated expressions
Previously, support rules generated for negated expressions did not
include arguments. Assuming the negated expression did not depend on
unknowns in the calling query this is fine. However, if that's not the
case, then the support rule must be able refer to the unknowns,
otherwise, the partial evaluation result will yield false positives.

The unknowns cannot be pushed down to the support rule via the binding
lists because the bindings do not support references, only vars and
values. To workaround this, we generate support rules that include
arguments which can be used to push the unknwons down.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>