Rewrite with modifiers to allow refs as values
These changes update the compiler to rewrite with modifiers to allow
refs or other terms that require evaluation as values. For example,
before this change, the following would not be allowed:

allow with input as data.test.some_test_input

Because the with value (data.test.some_test_input) requires evaluation.

After this change, the expression will be rewritten so that the with
value is bound to a variable earlier in the query and then substituted
with a variable.

This change will make tests cleaner as users will not have to bind
inputs to variables explicitly inside their test cases.

Fixes #701

Signed-off-by: Torin Sandall <torinsandall@gmail.com>