Refactor partial eval to use rule index
This change updates partial eval to use the rule index instead of
directly gathering all of the rules off the compiler. In addition to
removing the duplication, this change also allows for the indexer to be
used at enforcement-time in data filtering use cases where clients are
asking for the conditions instead of just decisions.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>