Fix variable namespacing in partial eval
These changes fix two issues related to variable namespacing during
partial eval.

The first issue was that refs rooted with local vars (e.g., x.foo) were
being plugged before evalTerm was called. As a result, if the ref head
was unknown, the expression would be saved with an incorrect binding
list. This is fixed by subsituting the ref head but preserving the
bindings instead of plugging. This way the correct binding list is
included with the saved expression.

The second issue was that ref heads were not being plugged correctly.
Ref heads are only plugged during partial eval, so this did not cause
problems elsewhere. The problem was that the binding list was not being
applied before the ref head was namespaced.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>