Update CI to use benchstat
These changes update the CI to use benchstat instead of running make
perf on the last N commits. With these changes, CI will run the
benchmarks against the tip and the target branch and then report the
results with benchstat. Eventually we can send the result of benchstat
back to the PR.

For benchstat runs, we have chosen a test count of 5. This is pretty
much arbitrary. We may need to increase this depending on how Travis
performs.

Since the PACKAGES variable is going away, the old coverage targets have
been removed. They were unused for quite a while. We can add them back
in the future if we want.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>