Make default decision path configurable
These changes just make the default policy decision path configurable.
This path is used when callers execute unversioned POST requests against
OPA's HTTP API. Library integrations are free to implement behaviour.
Relying on this kind of behaviour simplifies integration because
services do not have to be configured with a path to query.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>