Fix with value rewriting for call terms
If a call term was used as a with value, e.g., "p with input as f(1)"
the type checker would panic because the call was not expanded as
expected. These changes update the rewrite term expr stage to rewrite
with values like other AST nodes such as calls and composites.

Also, add an extra test case for with value rewriting to cover #772.

Fixes #916

Signed-off-by: Torin Sandall <torinsandall@gmail.com>