Refactor accounting of bindings
Previously we kept track of the number of times a binding was
substituted to determine whether it had to be added back into the query
result to ensure the ref was defined. These changes refactor that logic
to simply walk the result and check if the binding value is contained.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>