Inline negated expressions during partial eval
Previously partial eval would generate support rules for the result of
negated expressions. With these changes, the result of partially
evaluating a negated expression may be inlined into the containing
query.

The result of partially evaluating a negated expression may be inlined
if:

- The query set resulting partial eval does not contain any positive,
  non-ground expressions. E.g., p[x] cannot be safely negated so it
  would still result in support rules.

- The size of the cross product that represents the inlined result is
  not larger than 16. This limit is arbitrary and in the future we can
  make it configurable as needed.

In the future it may be possible to make the inlining decision more
sophisticated. For now though, because copy propagation is applied to
the result, it's often possible to inline results (since all of the
variables are removed.)

With these changes we plug the partial eval results and apply copy
propagation DURING partial eval instead of at the very end. This avoids
the need for extra stage after copy propagation that would walk the
result set, inline negation, and then cleanup the queries and support
modules. While the latter approach would allow for global optimizations,
it would be more complicated and not be any better for now (since we
are not applying global optimization.) To support adding plugged
expressions to the save stack, we have updated the save stack to keep
track of whether a plug operation has been performed (if the bindings
are nil then it's assumed that the contents have been plugged.) This
prevents the caller from accidentally calling plug twice (which would
cause namespacing issues.)

As part of these changes, we have modified the union-find structure in
the copy propagation implementation to rank live vars higher than
candidates for removal. This avoids adding extra vars to the results
(which helps the odds of being able to inline.)

Signed-off-by: Torin Sandall <torinsandall@gmail.com>