Fix unsafe support calls
The inlining changes introduced a bug in the support calls that are made
when the negation cannot be inlined. The support rule was being created
with one argument for each unknown in the entire save set. Since some of
those variables may not be preserved in the final result, it was
possible for the support call to become unsafe. To resolve this we only
add call args for the intersection of unknowns and vars that appear in
the result of partially evaluating the negated expression.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>