.travis.yml: set go_import_path, use deploy feature
When pushing a PR to _my_ remote (github.com/ME), travis will send me an
email saying that it failed to build, and keeps failing, for every
update of the branch[1].

Setting this go_import_path properly should fix the build issue.

Furthermore, having successfully tested the package, our .travis.yml
used to check if this was a PR (it's not, in the context of a branch
pushed to a fork), and did some docker image related tasks. This is done
using the deployment feature of travis now.[2]

[1]: https://travis-ci.com/srenatus/opa/builds/86468169
[2]: https://docs.travis-ci.com/user/deployment

Signed-off-by: Stephan Renatus <srenatus@chef.io>