Update wasm-test to build wasm library
Previously the wasm-test make target did not build the wasm libary and
update the checked in version. This was a bit clunky if you were making
changes to the wasm library implementation and wanting to run the
integration tests. It wasn't an issue before because we were not
extending the wasm library implementation much after the initial work.

With these changes, the wasm-test make target depends on the generate
target which will cause the wasm library to be rebuilt. The build
caching works so this doesn't slow down the feedback loop much.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>