Support for applying the with keyword to the data document
These changes make it possible to replace the data document.
Both base and virtual documents can be replaced. These changes support
replacing rules without arguments. They do not support replacing
rules/functions with arguments. To support that, we would need to take into
account the scenarios that would arise as a result of replacing the arguments
to the rule/function and the return value of the rule/function itself.

Fixes #517

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>