Fix release patch script to support LAST_VERSION
Previously the release script always computed LAST_VERSION from tags. We
recently published 0.9.3-rc tags for test purposes. This was preventing
thes cript generating the patch for the markdown/yaml files in the docs
that refer to the OPA version. These changes allow the caller to control
the LAST_VERSION so the docs get updated accordingly.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>