Fix out of range errors for eq/assign in compiler
The compiler was assuming that unification and assignment statements
were called with the correct number of operands. This is usually
enforced at parse-time by the grammar but if a unification or assignment
operation is called in prefix notation then it's possible for an
incorrect number of operands to be supplied (e.g., eq() or assign(x)).

These changes update the stages before type checking to tolerate these
kinds of invalid statements.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>