wasm: Add checks on single term and dot stmts
Previously we weren't checking whether the result of dot and single term
expressions were defined/not false. As a result, a query like input.x
where input is {} would be defined (when it should not.) Similarly, a
query like input.x where input is {"x": false} would be defined (when it
should not.)

These changes modify the planner and the backend to handle single term
expressions and dot statements respectively.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>