Fix semantic check for function references
These changes fix an issue in the semantic checks against function
references. The compiler is supposed to catch references to functions
that have not been applied (e.g., f(1) is OK but just f is not.) The
compiler was failing to catch these references when they were embedded
within composites. If these references were encountered during
evaluation, it resulted in undefined behaviour.

The fix for this issue is to simply remove the flag that gated whether
terms were checked. The flag was in place to deal with nested terms like
f(g(x)) but that has not been applicable for a while since these terms
are rewritten (and have been since the unifier was reimplemented over
year ago.)

Fixes #1132

Signed-off-by: Torin Sandall <torinsandall@gmail.com>