Fix parser to avoid duplicate comments in AST
The parser was accumulating comments in a parser-global slice. If the
parser backtracked, the comments were not thrown away. Because most
successful parses seem to backtrack, this meant that most ASTs contained
duplicated comments.

This change simply modifies the parser to accumulate the comments in a
set that just gets sorted at the end.

Fixes #426

Signed-off-by: Torin Sandall <torinsandall@gmail.com>