wasm: Rename ir.LoopStmt to ir.ScanStmt
The original ir.LoopStmt was essentially performing a scan. The
condition was not being used to break out of the loop. When we add
unification of composites, we're going to need an actual Loop
operation that can be halted, so renaming LoopStmt to ScanStmt to
better reflect it's implementation and also freeing up the name.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>