Add documentation in the REST API section for error 405
Fixes #1181

There are a couple of things reported in this PR: lack of 405 error and
missing return codes.

I decided to go through the entire REST API manual while comparing with
the code and found more discrepancies. These are the changes I've done:

1 - Restructured the sections to read

..rest api spec
..example

as opposed to:

..part of rest api spec
..example
..part of rest api spec.

As I went through the manual this break in the spec made things
confusing. If I was going to implement a client I would prefer to have
the complete spec for a certain method followed by an example as
opposed to the example break.

2 - Fixed missing or inconsistent return codes

There were quite a few of these.

3 - Added an overall 405 return code.

Since the sections are per method it does not make sense to have a 405
return code there. Therefore I added a small overall section in the
beginning that explains that we return 405 for non-allowed methods

Signed-off-by: repenno <rapenno@gmail.com>