internal: Add helper package to store manifests
Previously we had a bunch of ad-hoc code to store the revision and
roots internally. These changes just add a proper package to deal with
the reading and writing of the revision and roots.

Note, the bundle package itself is still hardcoding the manifest path
but we can fix that in a subsequent PR.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>