Build and tag rootless Docker image
Added a new `openpolicyagent/opa:latest-rootless` and
`openpolicyagent/opa:${version}-rootless` image that
does not run as root.

Did this instead of switching the existing image to
maintain backwards compatibility for people who may
be deploying the image as `root` and using the `root`
privileges to bind to privileged ports.

Signed-off-by: Geoff Baskwill <me@geoffbaskwill.ca>