Fix `net.cidr_contains` for partial match cases
The calculation for the last ip of a block had a typo switching the buffer
we created the last ip address from.. This meant we always returned the
first IP address in the range as the last one.

There was also a gap in the unit tests for these cases. Both are corrected
with this patch.

Fixes: #1303
Signed-off-by: Patrick East <east.patrick@gmail.com>