Add CLI config overrides and ENV injection
There are two new CLI options added with this change for the `opa run`
sub command.

`--set`
`--set-file`

These allow for overriding config options on the CLI using `key=value`
options to reference the YAML/JSON config structures. The `--set` value
expects to take in the value you want to set while the `--set-file` value is
a path to a file which will be read for the value of the file. This is primarily
useful for secrets mounted as files on a host.

In addition this adds in some simple environment variable injection so that a
deployer can specify environment variables in the config via ${NAME} syntax.
At the time the config is loaded from file it will inject in the environment vars.
This applies to strings set via the `--set` variable too.

Some things to note:
* This won't work for configs generated by discovery bundles, it is *only* for
  configurations loaded by the runtime.
* This is only done at the time of loading the file. Plugins receive modified
  copies of the config (post injection), and should never be allowed to re-read
  directly from disk.
* There are security implications of setting secrets in env vars. It is
  recommended to use the file based secrets approach with `--set-file`
  instead of environment variables.

Signed-off-by: Patrick East <east.patrick@gmail.com>