Add support for with keyword stacking (data)
This is a follow-on to the previous commit. Note, with this change,
input and data replacement are handled using the same code (whereas
before the data replacement used the baseCache which could lead to
slightly different behaviour). The only difference is that the data
replacement has to keep track of prefixes so that base/virtual
documents can be shadowed.

Also, rename helper that merges with values into term to indicate purpose.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>