Add Prepare() function to Rego objects
This adds in the `PrepareForEval` and `PrepareForPartial` functions to Rego objects
which lets a client Pre-parse and compile some parts of an Eval() for re-use.

These functions return a new PreparedXXQuery object that has specialized evaluation
Functions on them that can allow for changing some variables (input, unknowns, metrics,
and more) on the prepared query.

There is also a new metric for parsing in the input as this step is broken
out from the old workflow and decoupled from the query compiler.

Closes #1305

Signed-off-by: Patrick East <east.patrick@gmail.com>