Add an “edge” release to the docs
The “edge” version is the current tip of tree. This means for the live
site it is going to be the docs on master, for PR’s it will be the
current PR docs.

This leads us towards the same build/env for prod and dev/preview
builds of the docs.

Signed-off-by: Patrick East <east.patrick@gmail.com>