Get local netlify dev environment working
We needed to change to build the docs from the root of the git repo
Where the netlify toml file was. To make this easier we just shim the
docs-* make targets to the ./docs/Makefile.

The nice part of this is we get a full deployment with the redirects
and other netlify config bits in place.

Signed-off-by: Patrick East <east.patrick@gmail.com>