Add provenance query output
Add the new 'provenance' query parameter to the v1DataGet() and
v1DataPost() routines.

Set the fields of the DataResponseV1 struct with the version/build
values for possible return with the result of the data operations.

Add a unit test by setting version info, than ensuring we get the
strings in the response.  Note that the revision is left unset.

Signed-off-by: Peter W. Morreale <pmorreale@statestreet.com>