Fix extra ‘v’ in older versions of the docs
The older markdown still had `v{{< version >}}` shortcodes being used
which broke when the version changed to include the “v” in it.

This adds in some backwards compatibility logic for the shortcode and
adds newer, more specific, ones to replace it going forward. Part of
The issue is that typically the documentation will only want to
Reference the version that it is documenting and not always the
latest.

Fixes: #1382
Signed-off-by: Patrick East <east.patrick@gmail.com>