New --ignored-query-parameters option
Add a new runtime option: --ignored-query-parameters, that takes a
comma-separated list of OPA URL query parameters for OPA to ignore.

When set, OPA will return the normal response without the additional
data implied by the specific query parameters (assuming present on the
URL). The effect is identical to a REST API call without the specific
query parameter present.

Note that this setting is global - the entire OPA REST API honors this
option.

This allows administrators to directly control access to the data
implied by the query parameters directly from the start of an OPA instance.

The parameters, when encountered, are silently ignored.  If set,
the option string will appear in the OPA log.

Signed-off-by: Peter W. Morreale <pmorreale@statestreet.com>