ast: Fix set deep copy
The deep copy operation was not calling Copy on the set elements. This
manifested in issues in the REPL where parsed queries are retained and
used after the query compile runs. Since the query compile deep copies
the parsed query before mutating it, the REPL's parsed query was
getting corrupted.

Fixes #1406

Signed-off-by: Torin Sandall <torinsandall@gmail.com>