wasm: Refactor wasm backend to prepare for functions
These changes refactor the wasm backend so that we can emit
functions. The plan is compiled into a function so the same code can
be reused for other functions.

THe main change is to refactor most of the stages into common code
that will be reused by the plan and function compile stages.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>