wasm: Fix test runner to catch interrupts
These changes remove the -it flags from the docker run command used to
execute the Wasm tests. This will fix the original issue. In addition,
these changes update the test runner to catch interrupts so that
developers can still ctrl+c out of the test run.

Lastly, update the test generation program to create the directory
hierarchy instead of expecting the caller to do that.

Fixes #1431

Signed-off-by: Torin Sandall <torinsandall@gmail.com>