eval: Add parameter to control partial eval inlining
These changes add a new parameter to partial eval to disable inlining
on virtual documents. This allows callers with knowledge of the policy
to reduce the cost of partial eval (at the expense of receiving
support rules.) In the future, OPA can be extended to automatically
disable inlining using the new parameter.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>