Add documentation for calling OPA globally.
I added this documentation because I had to Google how to give
the OPA executable a global context so I can call the executable
from any path location. This documentation should be included
since it is only 2 lines and  new users won't have to use
Google to find the answer.

Signed-off-by: mbowen13 <mbowen@utexas.edu>