Fix parsing of var-like scalars
As reported in #1582, there are some issues with parsing variables that
start with `false`, `true` or `null`.

Two possible fixes for this issue are:

 1. Reordering the `Term` rule:

        Term <- Comprehension / Composite / Scalar / Call / Ref / Var

 2. Fixing the bug in `Scalar`: e.g. `nulls` should _not_ parse as a
    scalar by ensuring these keywords are not prefixes of a variable.
    As far as I know this is a fairly common technique in grammars.

It's bikeshedding the issue a bit but I would argue that (1) is not a
great solution: it really just works around the bug by moving things
around in `Term`: the bug may pop up again when reusing `Scalar` in
other places.

In order to do (2) nicely I also refactored the parser for variables
into one for var's first and subsequent characters.

Signed-off-by: Jasper Van der Jeugt <jasper@fugue.co>