rego: Support loading bundles and files w/ Rego API
There are three new APIs to load files, bundle files/dirs, or to
just give a bundle directly.

This requires re-shuffling some of the transaction handling for
a couple of reasons. First off, the bundle loading process requires
a write txn for placing any bundle manifets or data loaded into
the store. The second is that for auto-created txns we cant just
abort anymore, we need to commit after the prepare step.

So, we now conditionally create a write transaction and when we did
auto-create one we will commit it on success.

Signed-off-by: Patrick East <east.patrick@gmail.com>