build: Move fuzzer build into Docker image
The build was broken becasue the fuzzer artifacts were showing up in
the working copy and were owned by root once they got built inside of
a Docker image. This change updates the build to create an image for
the fuzzer to run inside of. The image can be shared across both
fuzzer runs and the artifacts are contained inside the image and don't
pollute the working copy.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>