bundle: Ensure data paths use `/` separators for key
Turns out on windows the filepath.Dir() call will give back paths that
are using `\` separators. This is problematic when we then go to split
the path with `/` to get the key. We need to make sure they are always
normalized to `/` separators.

Fixes: #1713
Signed-off-by: Patrick East <east.patrick@gmail.com>
(cherry picked from commit 13fab967f49271f788e255566aba622930033758)