wasm: Remove memory.grow calls on every malloc
The memory.grow calls were dominating the evaluation time and the
integration tests were beginning to take way too long to run. This
just updates the opa_malloc placeholder to keep track of the heap and
only call memory.grow when needed. We can revisit memory management in
the future but for now this is good enough.

Fixes #1121

Signed-off-by: Torin Sandall <torinsandall@gmail.com>