topdown: Base document dereference with composite
Previously this would raise an error when attempting to build the
storage path for some ref like `data.foo[["bar"]]`, but there isn't
really any reason why it should halt evaluation. It should just
be undefined.

This changes to look for this type of ref _before_ handing it off to
the store and will just return undefined now instead.

Fixes: #1057
Signed-off-by: Patrick East <east.patrick@gmail.com>