ast: Add map of rewritten vars to Compiler
The `ast.Compiler.RewrittenVars` variable will, after compilation,
give a map of generated vars to original vars unique to the compilers
module set.

This is useful for user facing strings where we can substitute the
original vars back in.

An important caveat for this is that there is no scoping information,
so this requires that generated variable names are unique for the
whole compiler.

This will only have the local variables that got rewritten.

Signed-off-by: Patrick East <east.patrick@gmail.com>