file/loader: Standardize on forward slash paths
When loading files we would build the list with, on windows, potential
paths with backslashes rather than forward slashes. This is
problematic because the check to trim the prefix for returning file
descriptors based on the root of the bundle.. On windows it could, in
some cases, fail to do so. This results in additional prefixes on the
loaded data.

Signed-off-by: Patrick East <east.patrick@gmail.com>