rego: Don't propagate non-threadsafe fields from Rego to preparedQuery
The metrics, tracers, and instrumentation are all not safe to be
shared from the Rego object to preparedQuery in cases where the
preparedQuery will be evaluated concurrently.

Callers of the API may pass the same one in to each (like what the
updated `opa eval` cmd does) if they know it is safe to do so. They
will now have to opt in to trace or instrument each part of preparing
and evaluating if using the separate API's.

This will affect anyone using preparedQueries and these fields if
they were not passing them in as `EvalOption`s before now.

Relates to: #1827 #1828
Signed-off-by: Patrick East <east.patrick@gmail.com>