Add debug tracing mode
This adds a new tracer mode for eval which will disable indexing so
that variable bindings for rules that would otherwise not evaluate
can be found.

This also adds metadata about the local variables which includes their
original text and source location. The JSON marshaller for locals will
now correctly show the key and value, plus added type information.

Fixes: #1697
Signed-off-by: Patrick East <east.patrick@gmail.com>