wasm: Fix bug in block nesting for scan statements
The backend was not nesting the scan loop inside of a block which
meant the emptiness check was breaking too far out. This meant that
statements following the scan were not being executed. This was only
affecting ordered rulesets because the defined statement that causes
short-circuiting was getting skipped and so we were falling through to
the next rule body and a conflict error would ultimately be thrown.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>