docs: Clarify kube-mgmt/Gatekeeper K8s docs
Previously the docs detailing the OPA-k8s integration only distinguished
between v1 (kube-mgmt) and v3 (gatekeeper) in the introduction page.
The debugging, primer, and tutorial pages did not explicitly state
which version they were detailing.

This change clarifies that those pages refer to the kube-mgmt
version and provide a pointer to the Gatekeeper readme.

Signed-off-by: Tim Hinrichs <tim@styra.com>