wasm: Extend wasm library to support shallow copying
The existing library functions return references and mutate values in
place (in case of append/insert/add operations). For with statements,
the shallow copy operation will be used to create copies that can be
safely modified without invalidating references.

In the future, we could investigate persistent versions of
array/set/object to avoid the need for these relatively expensive
copy operations.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>