wasm: Update tester to accept missing input
Previously the tester would supply a 'null' JSON value if one was not
defined in the asset file (because of how Go unmarshalling
works). This meant that input was always defined in the test
cases. This change updates the tester to accept missing input and
simply provide a NULL pointer to the eval() function which lets us
exercise cases where input is undefined.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>