ast: Fix module JSON unmarshalling
Previously when a module was unmarshalled from JSON the rules
contained in the module would not have their module pointers set. This
would cause the Path() function on the rule objects to panic.

This change updates the module struct to walk the rules and set the
module pointer accordingly.

This was discussed in #1780

Signed-off-by: Torin Sandall <torinsandall@gmail.com>