docs: Build html content from integrations.yaml
This adds in a the HTML view for the integrations.yaml and by having
hugo generate content at build time. It ends up being at the
`/docs/<version/ecosystem` url and there are links to it on the home
page, introduction page, and nav bar under the Misc category.

In future iterations we can add in sorting/filtering etc. It currently
just shows all of them.

Signed-off-by: Patrick East <east.patrick@gmail.com>