wasm: Update test framework to support skipping tests
Previously the generator would error out if the test case could not be
compiled. With this change, the generator just continues with the test
case marked as skipped. If the runner has verbose output enabled each
skip is reported. Once more built-in functions have been implemented
we can report skips by default.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>