wasm: Refactor with keyword implementation
This change fixes an issue in the with keyword implementation that
prevented the with statement from being used on negated
statements. The ir.WithStmt now includes a block that should be
executed with the modified local variable in scope. This simplifies
the planner's job (because it doesn't have to generate statements to
save modified locals) and makes the ir.WithStmt less error-prone. With
this change, it doesn't matter whether the planner processes negation
or with modifiers first.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>