topdown: Fix trace to unmangle variables
For the longest time, topdown traces have included rewritten/mangled
variable names. This makes it difficult for people to understand the
trace because the variable names they used in their policy do not show
up in the trace. Now that we have the rewritten variable map readily
available during evaluation we can easily provide the user-supplied
variable names.

Technically this commit just fixes the presentation problem. The
inclusion of the variable mapping was done in
d1522c647594264ef5996621d3ae3d0d739d19b1.

This change also updates the trace emitting code to include local
metadata for all vars that appear in the AST node instead of just vars
that have bindings. This allows us to unmangle all variables in the
node (e.g., otherwise it would not be possible unmangle Enter nodes.)

Fixes #1208

Signed-off-by: Torin Sandall <torinsandall@gmail.com>