docs: Fixup style for function signatures
We use the `opa-keep-it-together` class all over in the policy
reference page, but it wasn't actually defined anywhere. At some point
we must have lost it.

This adds back in a definition for it and configures things so that
the function signatures will no longer word wrap.

Going along with this is a tweak to the bulma content width by
slimming down the `$gap` size for `.container`'s. For more info see:
https://bulma.io/documentation/layout/container/

End result is that the doc content is now taking up more of the
available space on the screen which helps display the tables of
built-in functions a little bit better. We may want to consider a
different format for the in the future.

Signed-off-by: Patrick East <east.patrick@gmail.com>