docs/website: Fix live-blocks interaction on edge/live pages
This was changed as part of the homepage rework but ended up being
both unused and broken.. so reverting it back to the way it was
before.

The live blocks on the home page are not interactive, so this works
out fine as-is. In the future we may need to revisit.

Signed-off-by: Patrick East <east.patrick@gmail.com>