topdown: Add support for partial eval of with statements
This commit updates the evaluator to evaluate expressions including
with statements where the values of all with statements are
known. Any saves that occur in the process are modified to include the
appropriate with statements. In the future we can optimize this by
removing with statements that have no affect on the result.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>