ast: Do not index expressions containing with statements
The rule indexer was incorporating expressions that met the = or
glob.match requirements even if they had a with statement
applied. This caused false negatives for the index lookup result. In
practice, it's highly unlikely that users would write expressions that
could be indexed that also include with statements (e.g., `input.a = 1
with ...`) so it's unlikley that anyone is affected.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>