plugins/discovery: Add plugin status updates
The discovery plugin will update its status based on the first time
it initializes and activates configuration. If a bundle is configured
it will switch to OK the first time it activates the configuration of
that bundle. If no dynamic configuration is present it will just go
to an "ok" status at Start() time.

Signed-off-by: Patrick East <east.patrick@gmail.com>