topdown: Wrap all builtin functions for errors normalization
The older RegisterFunctionalBuiltinX functions would do this and it
gave the builtin error messaging/typing a nice uniform look and feel.

This changes the newer RegisterBuiltinFunc to do the same thing, so
any builtin will now either return a nil error or a `*topdown.Error`.
With an appropriate code, location, and message.

Fixes: #2101
Signed-off-by: Patrick East <east.patrick@gmail.com>