wasm: Memory export support for compiled policies.
Developer can decide whether the compiled policies are to import or
export their memory. This is for wasm runtimes not supporting memory
importing.

This patch takes the approach to modify the pre-built OPA binary at
the compilation time as necessary. The alternative would have been to
have two versions of the pre-built binary, one with memory import and
one with memory export.

The default is still to import the memory as before. That is, the
developer has to explicitly enable the exporting when using either go
build or rego package to compile policies to wasm.

Signed-off-by: Teemu Koponen <koponen@styra.com>