cmd/run: explain how to set input
I struggled with this for a while today until I stumbled across a brief
mention somewhere on the OPA website. I didn't see notes to set
repl.input anywhere else in "opa run --help" or in "help input" from the
REPL. Mentioning in the command help seemed more appropriate since "help
input" is contained in the repl package which may be used in other
contexts where the user doesn't have a way to provide repl.input.

Signed-off-by: Mark Rushakoff <mark.rushakoff@gmail.com>