cmd/fmt: Always write to stdout
There was an explicit check that would prevent writing to stdout if
the formatting didn't change anything.

This makes sense for the "write to file" or "diff" options as we can
save work by bailing out early, but if configured to write the file
contents to stdout we should do so in all cases.

Fixes: #2235
Signed-off-by: Patrick East <east.patrick@gmail.com>