topdown: Save comprehensions containing unknown terms
Earlier while partially evaluating comprehensions, we would namespace the variables in the comprehension and then save it. This would result in the mutation of the comprehension expression in the orginal query and as a result the same expression containing missing terms when re-encountered would not be PE-ed and hence produce incorrect results. This change does a deep copy of the comprehension before namespacing so that the original query is not modified.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>