topdown: fix `http.send` string conversions
In most cases, `http.send` was using `(*ast.Term).String()` to coerce
parameters to strings. However, this causes embedded special character
to be quoted, rather than using the actual string the caller provides.
The fix is to consolidate the string conversion and type check to ensure
that a string value was actually given.

Signed-off-by: James Peach <jpeach@vmware.com>