topdown: Fix for unsafe negated query in partial results
In copy propogation we ran into issues where an expression was negated
and required a variable in it to show up in a non-negated expression
to be safe. We had the binding for the var and could have created it
but there were two issues:

1) We didn't resolve the variable binding for refs in negated
   expressions. This left the "wrong" thing in the result.

2) Once we had the unified representation for the value in the result
   we needed to ensure that not only was the binding contained in the
   result somewhere, but it needed to appear in a non-negated
   expression.

The copy propagation code will then add an equality expression for
bindings that are left over and that are not safely contained in the
result. However, we don't always need the full equality expression.

If the key isn't found anywhere in the result we don't need the lhs of
the equality. We are essentially adding a variable that will be
ignored when all we care is that the rhs is defined. So now we will
only generate an expression for the rhs in those cases.

Fixes: #2045
Signed-off-by: Patrick East <east.patrick@gmail.com>