ast: Fix bug in output var analysis and add unit tests
There was a bug in the output var analysis whereby the implementation
would visit references inside closures when it should have
stopped. This was not seen because the output var analysis was
primarily used by the body safety check that special cases closures
and avoids thep problem. As part of the fix we now have a set of unit
tests for the functionality.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>