topdown/copypropagation: Rename "bindings" and use ValueMap
The older "bindings" name was making things more complicated as it
overlapped with the unionfind, which is also variable bindings..

What the "bindings" really were is a set of equality expressions which
had been removed from the query and needed to be accounted for later
on.

Just by renaming this it makes the code way easier to read through
and reason about.

Now that the unionfind can support more than just var's we can extend
the removed expressions to use any value as well. This doesn't
implement any of these changes yet, just setting the state for it in
the future.

Signed-off-by: Patrick East <east.patrick@gmail.com>