Update governance model to reflect areas of expertise
The previous governance model did not account for areas of expertise
like gatekeeper or other sub-projects. As the OPA project grows, it is
necessary to clarify responsibilities and ownership of different parts
of the project.

This commit updates the governance model to reflect (1) that we do not
hold votes on individual PRs (this was cargo-culted from an existing
CNCF document at the time and is not relevant) (2) organization voting
will be used to resolve conflicts and (3) maintainers statuses and
areas of expertise.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>