docs: Update philosophy page to discuss the document model
When we refreshed the docs last year we removed the page titled 'how
OPA works' and in the process lost the description of the OPA document
model that explains the base and virtual document concepts. Since then
a bunch of people have referred back to those docs after getting
started.

This change re-introduces the model and provides an explanation of
_why_ the model exists.

This change also replaces the infographic with text containing the
same content as that feels a bit more inline with the rest of the
page.

Fixes #2284

Signed-off-by: Torin Sandall <torinsandall@gmail.com>