topdown: http.send cache test use new test server
The older version of the cache testing for http.send didn't use
separate t.Run()'s for each case so it made it hard to see which case
was broken if the number of requests didn't match up with expected.
It also re-used the same test server which seemed to cause some
timing issues.

Signed-off-by: Patrick East <east.patrick@gmail.com>