ast: Set location text for rules when parsing
Previously we didn't have the correct rule location text (for the full
rule, head, value, else's, etc). This changes to set the text and adds
some tests to validate.

This affects the formatter somewhat with how it decides where to place
comments. The current output now more accurately reflects their
placement in the original policies.

There is one little hack to set the else head value locations, there
will need to be some more significant changes to the formatter before
we can change that.

Signed-off-by: Patrick East <east.patrick@gmail.com>