server: Fix data races around state modified on reload
This commit fixes another data race found by running go test
-race. The server was reading state modified by the trigger handler
without using a mutex or being inside of a transaction.

Note, we should think about how we can move away from relying on the
server transactions to implement critical sections in the server as
this is bound to bite us in the future.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>