wip: download: Remove ClearCache() from downloader interface
In the interest of keeping the downloader API as simple as possible,
just accept an error return value from the callback which the
downloader can decide what to do with. For now, the downloader will
only cache the etag if the download and the callback succeed. Note,
this is slightly different from before where the downloader would
reset the etag for some reason--it was not clear why this was
happening and returning to the known good etag seems like a fine solution.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>