topdown: Fix partial eval to save correct binding list
In dda04ef6cb241465c297faf2025bde20b2d0e881 we fixed a bug in PE that
caused unknown terms to get evaluated when they should have been
saved. However, the fix itself contained a bug inside evalTerm whereby
the evaluator would save the binding list from the query that produced
the unknown as opposed to the binding list for the reference suffix
that was being saved. This issue manifests in PE output with
references containing free variables instead of plugged values.

This commit fixes the issue by unifying the unknown value with a
generated variable associated with the reference suffix's binding
list. The reference to be saved is then constructed from the generated
variable. This ensures that all of the terms in the reference to be
saved are associated with the same, correct binding list.

Fixes #2368

Signed-off-by: Torin Sandall <torinsandall@gmail.com>