build: add author's GitHub id to changelog file
OPA uses the changelog file to recognize people who contribute to the
project by mentioning them in it. This commit introduces a way to fetch
the GitHub handle for the commit authors if they are not listed in the
MAINTAINERS.md file. It makes use of the GitHub API to do so, and thus
introduces an optional cli argument to pass the token to authenticate
requests, without which you might run into rate limit errors if you run
the script a couple of times.

Fixes #2323
Signed-off-by: Aditya <aditya10699@gmail.com>