docs: Merge cheat sheet content into reference
This change merges the cheat sheet content into the reference page--so
that there is one place that users can go to answer basic questions
about how to do X inside of Rego. Also, reorder the policy language
pages to go from language definitions => reference => testing =>
performance as this is the order we want users to navigate.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>