cmd: Fix shutdown race in "run" unit tests
The server was configured with a 0 second graceful shutdown period
which would periodically cause the shutdown to return an error.

This is corrected to have a 1 second shutdown period (plenty of time
as the tests should _NOT_ have any open client requests when they
finish, if they do they _should_ fail).

In addition we don't let the runtime do an os.Exit(1) on error, we
use the lower level `runtime.Serve()` API and catch the error in the
tests.

Signed-off-by: Patrick East <east.patrick@gmail.com>