topdown: Fix bug in set/array comprehension indexing
In 00a71ef465ffed82c5dcbd6ea141fff434c15c85 we added support for
comprehension indexing, however, there was a bug in the implementation
for sets and objects: the child query was closing over the
parent--we only had test coverage against arrays so this went
uncaught. This meant that sets and objects would still see O(N^2) runtime.

This commit resolves the issue and includes a new instrumentation
counter that should help identify this kind of issue (if the skip and
miss count differs, something is wrong.) Rather than erroring in this
case, we just lazily compute the comprehension.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>