ast: Fix locations on rules parsed from exprs
The location info on rules parsed from exprs was not ideal--it was
being set to the value of the rhs term (which is not accurate.)
This commit updates the parser helper function to reset the location
to the original expr location.

This commit also fixes an issue in the parser helper test case that
was introduced in 49a963f16acab85ac51ab7326fb53f4d742e711f when we
used a tool to roundtrip the policy examples in the OPA codebase
to move away from the old datalog if-then syntax (":-").

Signed-off-by: Torin Sandall <torinsandall@gmail.com>