format: Fix panic with else blocks and comments
When an else block had a comment in between it and the rule body we
were not starting a new line, which triggers a panic in the formatter.

This corrects the issue, and also removes the panics. If the functions
are not called correctly the formatting will be off and tests will
fail. There is little benefit in them causing a panic.

Fixes: #2420
Signed-off-by: Patrick East <east.patrick@gmail.com>