topdown: Add QueryTracer and deprecate Tracer
This renames a recently introduces (but not released)
`topdown.CustomTracer` to now be a `topdown.QueryTracer`. A new API is
added to replace the older `Trace()` API. Making this change has a
substantial impact on performance by eliminating allocations.

The `topdown.Query` and `topdown.eval` will now exclusively use the
new interface, but a compatibility layer is provided to wrap older
`topdown.Tracer` implementations with a shim `topdown.QueryTracer`
to provide transparent backwards compatibility for existing
integrations. In addition if a newer `QueryTracer` is passed into the
older `WithTracer()` API it will be "upgraded" to the newer type
without needing to be wrapped.

The `topdown.BufferTracer` is updated to support the new interface.

** Backwards Compatibility Changes: **

* `topdown.Tracer` has been deprecated in favor of a newer interface
  `topdown.QueryTracer`.
* All tracers (regardless of interface implementation) will now only
  be checked for being enabled at the beginning of query evaluation
  rather than on a per-event basis.
* `topdown.BuiltinContext#Tracers` has been deprecated in favor of
  `topdown.BuiltinContext#QueryTracers`. The older `Tracers` field
  will be `nil` starting this release, and eventually removed.

Signed-off-by: Patrick East <east.patrick@gmail.com>