wasm: Fix makefile to generate dependencies for all .c files
Previously the makefile did not include dependencies on header
files. If the header files changed the corresponding .wasm files would
not get rebuilt. This commit fixes the makefile to use clang's -MD
flag to generate makefile targets and then includes the generated
targets in the makefile. The -MP flag is used to handle file deletes
and renames.

This appears to be the best modern solution to the problem. For
details see: http://www.microhowto.info/howto/automatically_generate_makefile_dependencies.html

Signed-off-by: Torin Sandall <torinsandall@gmail.com>