wasm: Fix compiler to compute lowest free data segment offset
Previously the compiler would setup the data segment for
planner/compiler strings to initialize at offset 2048 in the linear
memory. With this change, the compiler finds the lowest free offset by
looking at the other data segments in the module. This fixes an issue
where the data segments from the wasm library were being overwritten.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>