wasm: opa_free implementation.
This transforms the malloc implementation to a simple first-fit
malloc. It maintains free memory blocks in a linked list, merging
adjacent blocks when possible. If no large enough free block is found
at the allocation time, a new one is allocated from heap ptr, just as
before. If the last free block ends to heap ptr, the block is removed
and heap pointer is equally subtracted. This is both to maintain the
existing heap ptr / top interface and yet to allow the heap to shrink.

Signed-off-by: Teemu Koponen <koponen@styra.com>