topdown: Add RuleTracer
The new RuleTracer implements a `topdown.QueryTracer` interface and
will filter on rule events counting enters, exits, and fails for all
rules encountered.

The VirtualDoc test has been updated to use it to give some idea on
the overhead involved. The intention is that this tracer will be
"always on".

Signed-off-by: Patrick East <east.patrick@gmail.com>