topdown: Fix support rule name suffix for negated expressions
Previously OPA was only using the outer query ID and expression index
to namespace support rules. However, this was incorrect if the same
expression was evaluated multiple times due to iteration. This commit
fixes the issue by including the query ID of the complemented
expression in the rule name suffix.

Fixes #2491

Signed-off-by: Torin Sandall <torinsandall@gmail.com>