Refactor fuzzit dockerfile and make targets
This makes them faster and (IMO) easier to understand what is
happening. The dockerfile now builds an image that has the required
tools, and the script (run in a container using the docker image) is
doing the actual building and job kickoff.

In the future we could split the image off more and publish it
separately, for now it just needs to be built locally for each CI job.

Signed-off-by: Patrick East <east.patrick@gmail.com>