wasm: Refactor built-in function name mapping
Previously, the names of built-in functions implemented in the
wasm library were mapped to the wasm library function names inside the
planner. This commit just refactors teh mapping so that it's defined
inside the wasm backend. The latter is where it really belongs since
(in theory) the planner could be used with other backends in the
future. As part of this change, some of the internal functions in the
backend have been renamed for clarity.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>