topdown: Mark vars as unknown in exprs saved due to with stmts
Previously, if an expression was saved due to a `with` statement
containing an unknown value, the vars in the expression were not
marked as unknown. This resulted in subsequent expressions that
depended on those unknowns being evaluated incorrectly.

This commit fixes the issue so that vars in the outputs from the
expression are marked as unknown.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>