ast: Remove hashing in object iterations.
This optimization does increase the memory usage slightly, though, and
in the long term, it is worth considering whether the iteration order
could become entirely random from callers perspective. It would allow
direct iteration over the elems without maintaining a separate values
slice.

The existing benchmark results do not degrade as a result of this, but
a particular large policy evaluation got 9% faster.

Signed-off-by: Teemu Koponen <koponen@styra.com>