ast: Fix panic in local var rewriting caused by object corruption
The local variable rewriting stage was mutating object keys while
iterating over the object itself. This caused the object to become
corrupted. This commit fixes the issue by switching the stage to use
the transformer interface which can safely mutate objects.

The transformer is more expensive because it preemptively copies
terms. If this is an issue, we could special case the transformer
usage to when an object would be mutated.

Fixes #2661

Signed-off-by: Torin Sandall <torinsandall@gmail.com>