ast: Fix parser to ignore rules with args and key in head
The parser was accepting rules that contained args and key terms in
the head. This was causing a panic during type checking because later
stages do expect this. This change just updates the parser to not emit
a rule in this case. Note, the parser was already dealing with this
correctly when the body was omitted.

Fixes #2662

Signed-off-by: Torin Sandall <torinsandall@gmail.com>