runtime: Fix data race in runtime package test
The runtime tests for the telemetry reporting feature contained a data
race because they were setting the version.Version variable which is
read by the runtime in other test cases. Since we now have the
version.Version value set in the source code, we do not have to
overwrite it in the test implementation.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>