plugins/logs: Correct hang in unit test fixture
Periodically we would get hangs in unit tests because the testServer
http handler would be stuck trying to send an event through the
channel. If the unit test was completed nothing would be reading,
and the send would be blocked indefinitely... as it turns out, the
http test server cannot shutdown while handlers are still running,
so the unit tests hang indefinitely.

To prevent this from happening we will drain any pending events when
stopping the testServer.

Signed-off-by: Patrick East <east.patrick@gmail.com>