topdown: Fix merge to copy input values if needed
This commit updates the merging logic to copy input values if
needed to prevent the input values from being mutated unexpectedly. A
copy is required if subsequent with statements would modify a previous
value--this can be caught by checking if any of the subsequent with
statement targets are prefixed by the current with statement target.

This change is required because although topdown plugs with statement
values before merging, the values are not deep copied--they are passed
by reference.

Fixes #2813

Signed-off-by: Torin Sandall <torinsandall@gmail.com>