wasm: Cache for built-ins' state.
Builtins can maintain state across their invocations, during a single
evaluation run. While we could leverage the builtin ctx parameter for
this, that context is always going to be evaluation run specific and
relying on that would not allow for eventual extension of the caching
across evaluations, which is what golang topdown builtins can do.

Signed-off-by: Teemu Koponen <koponen@styra.com>