test/cases: Specify sort_bindings on cases that require it
During integration testing of the golang-opa-wasm library, we found
that some of the exported test cases did not have the 'sort_bindings'
option set. This was because 'sort_bindings' was only set for cases
that generated an output value with a partial set--cases that
generated an output value using a set comprehension or array
comprehension would not have 'sort_bindings' set. This caused false
positive test failures in the golang-opa-wasm SDK tests.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>