plugins/bundle: Support for loading signed bundles from disk
Earlier when loading bundles from disk, we weren't providing
the bundle verification config that would be needed to read
a signed bundle that may have been persisted to disk in the past.

This changes adds support for loading signed bundles from disk by
providing the bundle verification config (if any) to the bundle
reader.

Fixes: #2824

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>