plugins/bundle: Fix default size limit
For the deprecated `bundle` config style bundle configuration we were
not setting the default bundle size limit, instead a zero value was
being propagated through when we "upgraded" the older style config
to the newer values.

This corrects that issue by always setting it to the default size
limit, which maintains the previous behavior.

Signed-off-by: Patrick East <east.patrick@gmail.com>