internal/wasm/sdk: Fix race in http loader shutdown
We would potentially set the `closing` channel to nil while a
goroutine was still attempting to read from it. This change corrects
the ordering so that the goroutine will catch when we start to close
and would trigger the for-loop in poller() to stop by cancelling the
context, rather than the select in the loop watching the `closing`
channel directly. This ensures that the context is cancelled before
returning, and that there isn't a race on the readers for `closing`.

Fixes: #2884

Signed-off-by: Patrick East <east.patrick@gmail.com>