wasm: Memoize planned functions without positional args
This commit updates the C library to expose a global key-value mapping
that can be initialized and supports push/pop operations for
shadowing. The key-value mapping is implemented using opa_object_t to
keep things simple. In the future, this could be replaced with
something more efficient.

The wasm compiler uses the key-value mapping to memoize calls to
planned functions that only depend on input and data. The compiler
uses the function index as the key and the function return value as
the value. The choice to memoize at the call-site as opposed to inside
planned functions was arbitrary.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>