internal/wasm/sdk: Lock for all data updates
Using some course locking for all 3 public API's on the pool which
can change external data on the pool vms. While a potential
performance bottleneck, we currently don't have any usage where we
would expect to call these methods in parallel, so it isn't expected
to be an issue right now. In the future we can potentially relax
these and use a more sophisticated system... but for now this will do.

Signed-off-by: Patrick East <east.patrick@gmail.com>