internal/wasm/sdk: Correct vm update behavior on error
There are two related issues being addressed. First, as we iterate
over the pools vms and update them, if any have an error we remove
that index. This process swaps the last one on the list with that
index and then shrinks the slice by 1. We were not copying the
"acquired" value for the last vm in the list to the removed index so
we could get out of sync.

The second issue is that when we removed a vm at some index we would
increment and continue updating at the next index. This meant that
the vm that was last on the list, which was just swapped in for the
removed index, would get skipped.

Signed-off-by: Patrick East <east.patrick@gmail.com>