internal/wasm/sdk: Run exported topdown tests
This adds in a new test case which will load the "external" test cases
from our topdown tests, and run them with the Wasm SDK.

There are a few adjustments made to get the tests passing, mostly due
to serializing the Rego types (sets) through JSON values. In addition
a bunch of tests have an exception for them, and associated github
issues filed. If/when we fix those we can remove them from the
`exceptions.yaml` file and have those test cases run.

Signed-off-by: Patrick East <east.patrick@gmail.com>