wasm: build without --export-all, annotate exports and "used" functions
This introduces three #define statements,

- WASM_EXPORT(name)
- OPA_BUILTIN
- OPA_INTERNAL

The first one makes LLVM _export_ a function (by the passed name), the
latter two make LLVM keep them, so they're not removed when it's deleting
dead code.

Under the hood, there is no distinction between OPA_BUILTIN and
OPA_INTERNAL. OPA_BUILTIN is used for functions that are the;
implementation of a builtin; OPA_INTERNAL is used for functions that
the planner generates calls to (i.e., parts implemented in C, but not
directly corresponding to builtins).

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>