ci: run npm-opa-wasm example
This attempts to give us some more safety for WASM-related changes.
We don't want to break the SDK!

It's introducing some coordination efforts in the future: when a WASM-
related change is supposed to be merged that requires SDKs to change,
we'll have to merge that into npm-opa-wasm first, before the PR can go
in. However, it's better to have this sort of breaking change smoke
test.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>