ast: use Transformer in rewriteEquals
Before, the WalkExprs modifying its operand would leave the object in
a bad state -- the key's Value was mutated, changing its hash, but the
object's hashtable wouldn't be updated.

Using the transformer, we'll get a new object with the transformation
applied to its keys and values -- it's more expensive but safe.

Fixes #2915.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>