compiler: emit debug messages for indexing decisions
Why a comprehension is not indexed should now be a little
more transparent, when enabling debug output:

    $ opa build --debug test.rego
    compile.go:1832: test.rego:3: no index vars

While changing this, the debug mechanism got more stream-y:
the different WithDebug() options take io.Writer, and they're
passed along to different sub-components. Nothing is retrieved
after (for example) compiling, but during compilation, the
debug messages are written to the passed writer.

The debug logs for optimization and planning now also include
the source file locations (go) where a debug message was logged.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>