wasm: Fix memory leak by using caller to obtain exports
There was a memory leak in the VM that was caused by circular
dependencies on the objects shared between Go and Rust. The circular
dependencies existed because the functions used for the built-ins and
opa_abort were closing over wasmtime.Store-related objects.

By using the Caller to obtain the exports, we can break the circular
dependency and avoid the memory leak.

This commit relies on the fix for
https://github.com/bytecodealliance/wasmtime-go/pull/59 which is in
the development version of wasmtime{-go} that is vendored in a
previous commit.

This change bumps the ABI minor version (1.0 => 1.1)

Signed-off-by: Torin Sandall <torinsandall@gmail.com>