Add SDK embryo
Create simple constructs for instantiating OPA with the same configuation
options as provided the standalone application.

Included in PR also a simple management API server implementation for testing,
which could be useful both for SDK integration testing as well as for internal
tests. Plan is to extend this and move much of the repeated logic around this
in plugin tests et. al. in here.

Closes #3140

Signed-off-by: Anders Eknert <anders@eknert.com>