Re-arrange structs for optimal size
This saves very little in terms of memory consumtion (8 bytes for most structs)
but on the other hand doesn't cost anything either. The only struct where I
think this will make a marginal difference is the Expr one since there can be
quite a few of those in a module.

Using https://github.com/orijtech/structslop

Signed-off-by: Anders Eknert <anders@eknert.com>