compiler/wasm: memoize in function (not callsite)
As soon as a function is called twice, there should be a little bit
to gain from having the memoization happen in the function body.

Run on a big bundle, this results in the following change in opcode
counts:

    Total opcodes: 184881 -> 183719

    Opcode counts:
    local.get: 46920 -> 46760 -
    i32.const: 24152 -> 23706 -
    local.set: 14097 -> 14097 =
    call:      12547 -> 12061 -
    end:       10911 -> 11015 +
    br_if:     11062 -> 10826 -
    block:      8987 ->  9092 +

So, the change isn't dramatic at all. But less is more, so I guess
we could still do this.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>