topdown: Update the logic to calculate the size of cache item
Previously when inserting an item in the inter-query cache, only
the size of the raw string representing the HTTP response message body
was used to calculate the amount of memory used by the item. This was
errornous as the actual item inserted in the cache would hold not only the
raw string but also the json representation of the message body.
As a result, the cache would end up consuming more memory
than the user-defined cache size limit. This change now returns a more accurate
measure of the cache item size.

Fixes: #3024

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>