wasm: experiment with proxy-wasm
We're using proxy-wasm-cpp-sdk, but without the protobuf bits.

To keep the exports, we'll set --export-dynamic on wasm-ld,
it'll blow our wasm modules up a bit, but bridging the emscripten
mechanism (attribute "used") with ours (attribute "export_name")
can be dealt with some other day.

This introduces a git submodule for the sdk, so after checkout,
do

    git submodule update --init --recursive

Note that we don't use the LOG_DEBUG etc macros, as they'll give us

    wasm-ld-11: error: _obj/src/http_wasm_example.wasm: undefined symbol: std::__1::to_string(int)

for reasons I don't completely understand yet.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>