planner: forget variables from comprehensions
Before, variables introduced in a comprehension, like the `x` in

    { x | xs[x] }

would have been recorded in `p.vars`, and have an effect outside of
the scope of this comprehension, leading to wrong plans.

Now, we'll push a new frame for learning about variables we've not
known before, and reset to that state after planning the
comprehensions, thereby forgetting them.

Fixes #3325.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>