ci: Use the ci container to build binary for make image
The `make image` target was relying on the build-docker target to
build the binary that goes into docker image. The problem is that
build-docker runs on the host and can therefore produce a binary for
the wrong environment.

This change just updates the image target to rely on
ci-go-build-docker instead which runs go build inside of a docker
container and ensures the environment is correct.

Since the user is already trying to build a docker image, it seems
reasonable to require docker to build the binary (normally we try to
avoid this dependency for plain binary builds.)

Signed-off-by: Torin Sandall <torinsandall@gmail.com>