ast: Update checker to propagate annotation types during inference
This commit updates the type checker so that types inferred from
annotations are propagated into the type environment (see the test
case). Also, combine duplicated if-else condition and only wrap the
type environment once--there is no need to wrap for each annotation as
far as I can see.

This change works by wrapping the type environment exactly once for
each rule and then unwrapping it before preparing the update based on
the inferred types in the body.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>