wasm-e2e: fix test runner, adapt and include jsonpatch tests
The `sort_bindings` key now is interpreted a little differently now: it's
no longer sorting and comparing, but building two sets and compares them.

From the test author's perspective, nothing has changed, except that it
now actually compares them.

Fixes #2949.

* add new exceptions, work on comparison logic

Turns out there are five hidden failures uncovered by de-noop-ing sort_bindings.

- Two have been related to deeper nesting than our current sorting can account
  for,
- two seem to be genuine bugs in comparing objects in WASM,

and one could be fixed right away by changing the query to not lead to deep
nesting.

* exported tests: don't sort test expectations

This leads to very weird situations, and seems to only be a convenience
for the test authors. So, instead, we'll have the test authors pin down
whatever the sorted bindings are, and assert that the sorted bindings of
the returned result set match.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>