website: link to same page (if possible) from version picker
This might not be the most efficient or idiomatic way to do it, but
I think it works.

Doesn't do in-page anchors, so selecting a different version when on
page /docs/v0.26.0/policy-performance/#benchmarking-queries will get
you to /docs/<other>/policy-performance/ only.

Fixes #3023 to some extend. A little less annoying now, I hope.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>