wasm: remove opa_number_float, introduce opa_number_mpd
This changes the hash function used for numbers -- the implication
is that set ordering of sets of numbers changes.

Also, this includes a dirty fix applied in the vendored libmpdec
sources: we're using its string formatting function now, and that
would always print the sign of the exponent. We don't want that,
so it's changed.

Now, for anything that requires going through mpd, we'll do it
only once: opa_number_ref will parse the string, and return the
OPA_NUMBER_REPR_MPD. Retrieving that decimal is done via
opa_number_to_bf.

In some intermediate state, I had been convering all the integer
representations into mpd_t, too, but that turned out to be too
costly.

Now, to properly hash numbers (ie. '2' needs the same hash regardless
of whether its represenation is an mpd_t or an int), we'll convert
the integers to mpd_t in opa_number_hash. It's not beautiful, but
does the trick.

opa_number_to_bf will also return a mpd_t for numbers stored as
integers, and that'll need to be free'd. To make that happen, we
always mpd_free the result of opa_number_to_bf -- but for those
instances that have an mpd representation, that'll have the static
flags set, causing mpd_free to NOT free it.

opa_number_ref now expects the string to be \0-terminated. The json parsing
functions, we'll not make a copy to ensure that it _is_ terminated; in the
compiled-policy-calls to opa_number_ref, we've already ensured that when
writing the strings into data.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>