plugins/logs/plugin_test: avoid race, bump github.com/sirupsen/logrus
Some mutexes had to be locked/unlocked to avoid the races. This showed
another one, related to the use of hooks.NewLocal() of logrus.

To fix that other one, I've first tried updating logrus (there was a
mention of fixed races in the changelog), but to no avail. Setting up
the hook before any plugin would log from that test resolved the issue.

No harm in updating logrus, though, let's keep that: 1.6.0 -> 1.8.1

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>